Restrict Add-On Update Frequency
Submitting a new version for an add-on every few hours is decidedly hostile to anyone who tries to use the add-on seriously, and also a minor maintenance nightmare. So we politely "ask" overzealous scripters to tone it down a bit.

The interval for submitting updates for existing add-ons is now lower-bordered at 3 days. Attempts to update more often than that are rejected.